refactor(remote): separate GPU provider models and creation payload
